app/requirements.txt
cachelib==0.10.2
click==8.1.3
Flask==2.2.2
Flask-Session==0.4.0
greenlet==2.0.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
protobuf==3.20.3
pycodestyle==2.10.0
python-dotenv==1.0.0
SQLAlchemy==2.0.3
SQLAlchemy-Utils==0.40.0
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
